typescriptification de picture-index et bonne instantiation alpine-data
