+++ core/templates/core/user_pictures.jinja
@@ -6,7 +6,7 @@
 {%- endblock -%}
 
 {% block additional_js %}
-  <script type="module" src="{{ static('bundled/user/pictures-index.js') }}"></script>
+  <script type="module" src="{{ static('bundled/user/pictures-index.ts') }}"></script>
 {% endblock %}
 
 {% block title %}
